Change scope of scraper to only search pages
Due to JS loaded html on the job posting pages, more time and learning
is needed to effectively scrape the job skills on those pages. In order
to make use of the scraper now, position titles and dates have been
scraped from the job search pages instead of the job posting pages. This
gives useable data today while the other information will be added as a
feature at a later date.